Reorganize GtkBackend into a file per backend feature
