+++ Sources/GtkBackend/Features/Alerts.swift
@@ -1,0 +1,62 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.Alerts {
+    public typealias Alert = Gtk.MessageDialog
+
+    public func createAlert() -> Alert {
+        let dialog = Gtk.MessageDialog()
+
+        // Register a custom shortcut controller to disable the default Escape-to-close
+        // action. In future we'll probably want to conditionally re-enable this
+        // shortcut in scenarios where we know which action button is the cancel action.
+        let controller = gtk_shortcut_controller_new()
+        let trigger = gtk_shortcut_trigger_parse_string("Escape")
+        let action = gtk_callback_action_new({ _, _, _ in return 1 }, nil, { _ in })
+        let shortcut = gtk_shortcut_new(trigger, action)
+        gtk_shortcut_controller_add_shortcut(controller, shortcut)
+        gtk_widget_add_controller(dialog.widgetPointer, controller)
+
+        return dialog
+    }
+
+    public func updateAlert(
+        _ alert: Alert,
+        title: String,
+        actionLabels: [String],
+        environment: EnvironmentValues
+    ) {
+        alert.text = title
+        for (i, label) in actionLabels.enumerated() {
+            alert.addButton(label: label, responseId: i)
+        }
+    }
+
+    public func showAlert(
+        _ alert: Alert,
+        window: Window?,
+        responseHandler handleResponse: @escaping (Int) -> Void
+    ) {
+        alert.response = { _, responseId in
+            guard responseId != Int(UInt32(bitPattern: -4)) else {
+                // Ignore escape key for now. Once we support detecting
+                // the primary and secondary actions of alerts we can wire
+                // this up to whichever action is the default cancellation
+                // action.
+                return
+            }
+
+            alert.destroy()
+            handleResponse(responseId)
+        }
+        alert.isModal = true
+        alert.isDecorated = false
+        alert.setTransient(for: window ?? windows[0])
+        alert.show()
+    }
+
+    public func dismissAlert(_ alert: Alert, window: Window?) {
+        alert.destroy()
+    }
+}
