+++ Sources/GtkBackend/Features/Checkboxes.swift
@@ -1,0 +1,29 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.Checkboxes {
+    public func createCheckbox() -> Widget {
+        return Gtk.CheckButton()
+    }
+
+    public func updateCheckbox(
+        _ checkboxWidget: Widget,
+        environment: EnvironmentValues,
+        onChange: @escaping (Bool) -> Void
+    ) {
+        let checkboxWidget = checkboxWidget as! Gtk.CheckButton
+        checkboxWidget.sensitive = environment.isEnabled
+        checkboxWidget.notifyActive = { widget, _ in
+            onChange(widget.active)
+        }
+    }
+
+    public func setState(ofCheckbox checkboxWidget: Widget, to state: Bool) {
+        let checkboxWidget = checkboxWidget as! Gtk.CheckButton
+
+        checkboxWidget.withBlockedSignal(named: "notify::active") {
+            checkboxWidget.active = state
+        }
+    }
+}
