+++ Sources/GtkBackend/Features/Core.swift
@@ -1,0 +1,124 @@
+import CGtk
+import Foundation
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.Core {
+    public var deviceClass: DeviceClass { .desktop }
+
+    public func runMainLoop(_ callback: @escaping @MainActor () -> Void) {
+        gtkApp.run { window in
+            self.measurementCustomLabel = (self.createTextView() as! CustomLabel)
+            self.precreatedWindow = window
+            callback()
+
+            let provider = CSSProvider()
+            provider.loadCss(
+                from: """
+                    list {
+                        background: none;
+                    }
+
+                    list > row {
+                        padding: 0;
+                        min-height: 0;
+                    }
+
+                    .navigation-sidebar {
+                        margin: 0;
+                        padding: 0;
+                    }
+
+                    .navigation-sidebar > row { margin: 0;
+                        padding: 0;
+                    }
+                    """
+            )
+
+            // Keep a reference around so that the provider doesn't get removed as
+            // soon as we exit this scope.
+            self.globalCSSProvider = provider
+
+            #if !os(macOS)
+                Self.mainRunLoopTicklingLoop()
+            #endif
+        }
+    }
+
+    private static func mainRunLoopTicklingLoop(nextDelayMilliseconds: Int? = nil) {
+        Self.runInMainThread(afterMilliseconds: nextDelayMilliseconds ?? 50) {
+            // This performs one pass through the run loop
+            let nextDate = RunLoop.main.limitDate(forMode: .default)
+
+            // This isn't expected to be nil, but if it is we can just loop
+            // again quickly with the default delay.
+            let nextDelay = nextDate.map {
+                return max(min(Int($0.timeIntervalSinceNow * 1000), 50), 0)
+            }
+            mainRunLoopTicklingLoop(nextDelayMilliseconds: nextDelay)
+        }
+    }
+
+    public func runInMainThread(action: @escaping @MainActor () -> Void) {
+        let action = ThreadActionContext(action: action)
+        g_idle_add_full(
+            0,
+            { context in
+                guard let context else {
+                    fatalError("Gtk action callback called without context")
+                }
+
+                let action = Unmanaged<ThreadActionContext>.fromOpaque(context)
+                    .takeUnretainedValue()
+                let innerAction = action.action
+                MainActor.assumeIsolated {
+                    innerAction()
+                }
+
+                return 0
+            },
+            Unmanaged<ThreadActionContext>.passRetained(action).toOpaque(),
+            { _ in }
+        )
+    }
+
+    private static func runInMainThread(
+        afterMilliseconds delay: Int,
+        action: @escaping @MainActor () -> Void
+    ) {
+        let action = ThreadActionContext(action: action)
+        g_timeout_add_full(
+            0,
+            guint(max(delay, 0)),
+            { context in
+                guard let context else {
+                    fatalError("Gtk action callback called without context")
+                }
+
+                let action = Unmanaged<ThreadActionContext>.fromOpaque(context)
+                    .takeUnretainedValue()
+                let innerAction = action.action
+                MainActor.assumeIsolated {
+                    innerAction()
+                }
+
+                // Cancel the recurring timeout after one iteration
+                return 0
+            },
+            Unmanaged<ThreadActionContext>.passRetained(action).toOpaque(),
+            { _ in }
+        )
+    }
+
+    public func computeRootEnvironment(defaultEnvironment: EnvironmentValues) -> EnvironmentValues {
+        defaultEnvironment
+            .with(\.appPhase, windows.contains(where: \.isActive) ? .active : .inactive)
+    }
+
+    public func setRootEnvironmentChangeHandler(
+        to action: @escaping @Sendable @MainActor () -> Void
+    ) {
+        // TODO: React to theme changes
+        self.rootEnvironmentChangeHandler = action
+    }
+}
