+++ Sources/GtkBackend/Features/CornerRadius.swift
@@ -1,0 +1,13 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.CornerRadius {
+    public func createCornerRadiusContainer(wrapping child: Widget) -> Widget {
+        child
+    }
+
+    public func setCornerRadius(of widget: Widget, to radius: Int) {
+        widget.css.set(property: .cornerRadius(radius))
+    }
+}
