+++ Sources/GtkBackend/Features/ExternalURLs.swift
@@ -1,0 +1,11 @@
+import CGtk
+import Foundation
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.ExternalURLs {
+    public func openExternalURL(_ url: URL) throws {
+        // Used instead of gtk_uri_launcher_launch to maintain <4.10 compatibility
+        gtk_show_uri(nil, url.absoluteString, guint(GDK_CURRENT_TIME))
+    }
+}
