+++ Sources/GtkBackend/Features/GenericContainers.swift
@@ -1,0 +1,40 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.GenericContainers {
+    public func createContainer() -> Widget {
+        return Fixed()
+    }
+
+    public func removeAllChildren(of container: Widget) {
+        let container = container as! Fixed
+        container.removeAllChildren()
+    }
+
+    public func insert(_ child: Widget, into container: Widget, at index: Int) {
+        let container = container as! Fixed
+        container.put(child, index: index, x: 0, y: 0)
+    }
+
+    public func swap(childAt firstIndex: Int, withChildAt secondIndex: Int, in container: Widget) {
+        // Gtk.Fixed doesn't let us rearrange children, so we just swap them in
+        // our own list so that at least everything works on the SCUI side. The
+        // only side effect of this approach is that overlapping widgets may
+        // end up with unexpected z ordering. If that becomes an issue we may
+        // have to make a custom replacement for Gtk.Fixed.
+        let container = container as! Fixed
+        container.children.swapAt(firstIndex, secondIndex)
+    }
+
+    public func setPosition(ofChildAt index: Int, in container: Widget, to position: SIMD2<Int>) {
+        let container = container as! Fixed
+        container.move(container.children[index], x: Double(position.x), y: Double(position.y))
+    }
+
+    public func remove(childAt index: Int, from container: Widget) {
+        let container = container as! Fixed
+        let child = container.children[index]
+        container.remove(child)
+    }
+}
