+++ Sources/GtkBackend/Features/HoverGestures.swift
@@ -1,0 +1,28 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.HoverGestures {
+    public func createHoverTarget(wrapping child: Widget) -> Widget {
+        child.addEventController(EventControllerMotion())
+        return child
+    }
+
+    public func updateHoverTarget(
+        _ hoverTarget: Widget,
+        environment: EnvironmentValues,
+        action: @escaping (Bool) -> Void
+    ) {
+        let gesture =
+            hoverTarget.eventControllers.first { $0 is EventControllerMotion }
+                as! EventControllerMotion
+        gesture.enter = { _, _, _ in
+            guard environment.isEnabled else { return }
+            action(true)
+        }
+        gesture.leave = { _ in
+            guard environment.isEnabled else { return }
+            action(false)
+        }
+    }
+}
