+++ Sources/GtkBackend/Features/Images.swift
@@ -1,0 +1,48 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.Images {
+    public var requiresImageUpdateOnScaleFactorChange: Bool { false }
+
+    public func createImageView() -> Widget {
+        let imageView = Gtk.Picture()
+        imageView.keepAspectRatio = false
+        imageView.canShrink = true
+        return imageView
+    }
+
+    public func updateImageView(
+        _ imageView: Widget,
+        rgbaData: [UInt8],
+        width: Int,
+        height: Int,
+        targetWidth: Int,
+        targetHeight: Int,
+        dataHasChanged: Bool,
+        environment: EnvironmentValues
+    ) {
+        guard dataHasChanged else {
+            return
+        }
+
+        let imageView = imageView as! Gtk.Picture
+        let buffer = UnsafeMutableBufferPointer<UInt8>.allocate(capacity: rgbaData.count)
+        memcpy(buffer.baseAddress!, rgbaData, rgbaData.count)
+        let pixbuf = gdk_pixbuf_new_from_data(
+            buffer.baseAddress,
+            GDK_COLORSPACE_RGB,
+            1,
+            8,
+            Int32(width),
+            Int32(height),
+            Int32(width * 4),
+            { buffer, _ in
+                buffer?.deallocate()
+            },
+            nil
+        )
+        let texture = gdk_texture_new_for_pixbuf(pixbuf)!
+        imageView.setPaintable(texture)
+    }
+}
