+++ Sources/GtkBackend/Features/PopoverMenus.swift
@@ -1,0 +1,73 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.PopoverMenus {
+    public typealias Menu = Gtk.PopoverMenu
+
+    public func createPopoverMenu() -> PopoverMenu {
+        let menu = Gtk.PopoverMenu()
+        menu.hasArrow = false
+        return menu
+    }
+
+    public func updatePopoverMenu(
+        _ menu: Menu,
+        content: ResolvedMenu,
+        environment: EnvironmentValues
+    ) {
+        // Update menu model and action handlers
+        let actionGroup = Gtk.GSimpleActionGroup()
+        menu.model = renderMenu(
+            content,
+            actionMap: actionGroup,
+            actionNamespace: "menu",
+            actionPrefix: nil,
+            environment: environment
+        )
+        menu.insertActionGroup("menu", actionGroup)
+
+        // Compute styles
+        let menuBackground: Gtk.Color
+        let menuItemHoverBackground: Gtk.Color
+        let foreground = environment.suggestedForegroundColor.resolve(in: environment).gtkColor
+        switch environment.colorScheme {
+            case .light:
+                menuBackground = Gtk.Color(1, 1, 1)
+                menuItemHoverBackground = Gtk.Color(0.9, 0.9, 0.9)
+            case .dark:
+                menuBackground = Gtk.Color(0.175, 0.175, 0.175)
+                menuItemHoverBackground = Gtk.Color(1, 1, 1, 0.1)
+        }
+
+        // Set styles
+        menu.cssProvider.loadCss(
+            from: """
+                contents {
+                    background: \(CSSProperty.rgba(menuBackground));
+                }
+                contents modelbutton:hover, contents modelbutton:selected {
+                    background: \(CSSProperty.rgba(menuItemHoverBackground));
+                }
+                contents modelbutton label {
+                    color: \(CSSProperty.rgba(foreground));
+                }
+                contents modelbutton {
+                    color: \(CSSProperty.rgba(foreground));
+                }
+                """
+        )
+    }
+
+    public func showPopoverMenu(
+        _ menu: Menu,
+        at position: SIMD2<Int>,
+        relativeTo widget: Widget,
+        closeHandler handleClose: @escaping () -> Void
+    ) {
+        menu.popUpAtWidget(widget, relativePosition: position)
+        menu.onHide = {
+            handleClose()
+        }
+    }
+}
