+++ Sources/GtkBackend/Features/ProgressSpinners.swift
@@ -1,0 +1,11 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.ProgressSpinners {
+    public func createProgressSpinner() -> Widget {
+        let spinner = Spinner()
+        spinner.spinning = true
+        return spinner
+    }
+}
