+++ Sources/GtkBackend/Features/RevealFiles.swift
@@ -1,0 +1,43 @@
+import CGtk
+import Foundation
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.RevealFiles {
+    public func revealFile(_ url: URL) throws {
+        var success = false
+
+        #if !os(Windows)
+            let fileURI = url.absoluteString.replacingOccurrences(
+                of: ",",
+                with: "\\,"
+            )
+            let process = Process()
+            process.arguments = [
+                "dbus-send",
+                "--print-reply",
+                "--dest=org.freedesktop.FileManager1",
+                "/org/freedesktop/FileManager1",
+                "org.freedesktop.FileManager1.ShowItems",
+                "array:string:\(fileURI)",
+                "string:",
+            ]
+            process.executableURL = URL(fileURLWithPath: "/usr/bin/env")
+
+            do {
+                try process.run()
+                process.waitUntilExit()
+
+                success = process.terminationStatus == 0
+            } catch {
+
+            }
+        #endif
+
+        if !success {
+            // Fall back to opening the parent directory without highlighting
+            // the file.
+            try openExternalURL(url.deletingLastPathComponent())
+        }
+    }
+}
