+++ Sources/GtkBackend/Features/Sheets.swift
@@ -1,0 +1,121 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.Sheets {
+    public class Sheet: Gtk.Window {
+        var onDismiss: (() -> Void)? = nil
+        var interactiveDismissDisabled = false
+        var nestedSheet: Sheet?
+    }
+
+    static let defaultSheetCornerRadius = 10
+
+    public func createSheet(content: Widget) -> Sheet {
+        let sheet = Sheet()
+        sheet.setChild(content)
+
+        // Listen for interactive dismissals
+        sheet.onCloseRequest = { [weak self, weak sheet] _ in
+            guard let self, let sheet else {
+                return
+            }
+
+            self.runInMainThread {
+                self.dismissSheet(sheet)
+                sheet.onDismiss?()
+            }
+        }
+
+        // Allow the escape key to be used to dismiss interactively dismissible
+        // sheets.
+        sheet.setEscapeKeyPressedHandler { [weak self, weak sheet] in
+            guard let self, let sheet, !sheet.interactiveDismissDisabled else {
+                return
+            }
+
+            self.runInMainThread {
+                self.dismissSheet(sheet)
+                sheet.onDismiss?()
+            }
+        }
+
+        return sheet
+    }
+
+    public func updateSheet(
+        _ sheet: Sheet,
+        window: Window,
+        environment: EnvironmentValues,
+        size: SIMD2<Int>,
+        onDismiss: @escaping () -> Void,
+        cornerRadius: Double?,
+        detents: [PresentationDetent],
+        dragIndicatorVisibility: Visibility,
+        backgroundColor: SwiftCrossUI.Color.Resolved?,
+        interactiveDismissDisabled: Bool
+    ) {
+        sheet.size = Size(width: size.x, height: size.y)
+        sheet.onDismiss = onDismiss
+
+        // Add a slight border to not be just a flat corner
+        sheet.css.clear()
+        sheet.css.set(
+            property: .border(
+                color: SwiftCrossUI.Color.gray.resolve(in: environment).gtkColor,
+                width: 1
+            )
+        )
+
+        // Respect corner radius and background Color
+        let radius = cornerRadius.map(Int.init) ?? Self.defaultSheetCornerRadius
+        sheet.css.set(property: .cornerRadius(radius))
+        if let backgroundColor {
+            sheet.css.set(property: .backgroundColor(backgroundColor.gtkColor))
+        }
+
+        sheet.interactiveDismissDisabled = interactiveDismissDisabled
+
+        // - detents are only supported on mobile so we ignore them.
+        // - dragIndicatorVisibility is only supported on mobile so we ignore it.
+    }
+
+    public func presentSheet(_ sheet: Sheet, window: Window, parentSheet: Sheet?) {
+        let parent = parentSheet ?? window
+        sheet.isModal = true
+        sheet.isDecorated = false
+        sheet.destroyWithParent = true
+        if let parentSheet {
+            parentSheet.nestedSheet = sheet
+        }
+        sheet.setTransient(for: parent)
+        sheet.present()
+    }
+
+    public func dismissSheet(_ sheet: Sheet, window: Window, parentSheet: Sheet?) {
+        dismissSheet(sheet)
+        parentSheet?.nestedSheet = nil
+    }
+
+    private func dismissSheet(_ sheet: Sheet) {
+        // Dismiss the nested sheets from the topmost down. We could use
+        // recursion here, but then unbounded nested sheets would allow for
+        // users to cause programs to run out of stack relatively easily.
+        var nestedSheets: [Sheet] = []
+        var currentSheet = sheet
+        while let nestedSheet = currentSheet.nestedSheet {
+            nestedSheets.append(nestedSheet)
+            currentSheet = nestedSheet
+        }
+        for nestedSheet in nestedSheets.reversed() {
+            nestedSheet.destroy()
+            nestedSheet.onDismiss?()
+        }
+
+        sheet.destroy()
+    }
+
+    public func size(ofSheet sheet: Sheet) -> SIMD2<Int> {
+        return SIMD2(x: sheet.size.width, y: sheet.size.height)
+    }
+}
