+++ Sources/GtkBackend/Features/SplitViews.swift
@@ -1,0 +1,46 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.SplitViews {
+    public func createSplitView(leadingChild: Widget, trailingChild: Widget) -> Widget {
+        let widget = Paned(orientation: .horizontal)
+        let leadingContainer = wrapInCustomRootContainer(leadingChild)
+        let trailingContainer = wrapInCustomRootContainer(trailingChild)
+
+        widget.startChild = leadingContainer
+        widget.endChild = trailingContainer
+        widget.shrinkStartChild = false
+        widget.shrinkEndChild = false
+
+        widget.position = 200
+        return widget
+    }
+
+    public func setResizeHandler(
+        ofSplitView splitView: Widget,
+        to action: @escaping () -> Void
+    ) {
+        let splitView = splitView as! Paned
+        splitView.notifyPosition = { _ in
+            action()
+        }
+    }
+
+    public func sidebarWidth(ofSplitView splitView: Widget) -> Int {
+        let splitView = splitView as! Paned
+        return splitView.position
+    }
+
+    public func setSidebarWidthBounds(
+        ofSplitView splitView: Widget,
+        minimum minimumWidth: Int,
+        maximum maximumWidth: Int
+    ) {
+        let splitView = splitView as! Paned
+        show(widget: splitView.startChild!)
+        let width = splitView.getNaturalSize().width
+        splitView.startChild?.setSizeRequest(width: minimumWidth, height: 0)
+        splitView.endChild?.setSizeRequest(width: width - maximumWidth, height: 0)
+    }
+}
