+++ Sources/GtkBackend/Features/StringLabelButtons.swift
@@ -1,0 +1,24 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.StringLabelButtons {
+    public func createSimpleButton() -> Widget {
+        return Button()
+    }
+
+    public func updateSimpleButton(
+        _ button: Widget,
+        label: String,
+        environment: EnvironmentValues,
+        action: @escaping () -> Void
+    ) {
+        // TODO: Update button label color using environment
+        let button = button as! Gtk.Button
+        button.sensitive = environment.isEnabled
+        button.label = label
+        button.clicked = { _ in action() }
+        button.css.clear()
+        button.css.set(properties: Self.cssProperties(for: environment, isControl: true))
+    }
+}
