+++ Sources/GtkBackend/Features/Switches.swift
@@ -1,0 +1,29 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.Switches {
+    public func createSwitch() -> Widget {
+        return Switch()
+    }
+
+    public func updateSwitch(
+        _ switchWidget: Widget,
+        environment: EnvironmentValues,
+        onChange: @escaping (Bool) -> Void
+    ) {
+        let switchWidget = switchWidget as! Gtk.Switch
+        switchWidget.sensitive = environment.isEnabled
+        switchWidget.notifyActive = { widget, _ in
+            onChange(widget.active)
+        }
+    }
+
+    public func setState(ofSwitch switchWidget: Widget, to state: Bool) {
+        let switchWidget = switchWidget as! Gtk.Switch
+
+        switchWidget.withBlockedSignal(named: "notify::active") {
+            switchWidget.active = state
+        }
+    }
+}
