+++ Sources/GtkBackend/Features/TapGestures.swift
@@ -1,0 +1,65 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.TapGestures {
+    public func createTapGestureTarget(wrapping child: Widget, gesture: TapGesture) -> Widget {
+        var gtkGesture: GestureSingle
+        switch gesture.kind {
+            case .primary:
+                gtkGesture = GestureClick()
+            case .secondary:
+                gtkGesture = GestureClick()
+                gtk_gesture_single_set_button(gtkGesture.opaquePointer, guint(GDK_BUTTON_SECONDARY))
+            case .longPress:
+                gtkGesture = GestureLongPress()
+        }
+        child.addEventController(gtkGesture)
+        return child
+    }
+
+    public func updateTapGestureTarget(
+        _ tapGestureTarget: Widget,
+        gesture: TapGesture,
+        environment: EnvironmentValues,
+        action: @escaping () -> Void
+    ) {
+        switch gesture.kind {
+            case .primary:
+                let gesture =
+                    tapGestureTarget.eventControllers.first {
+                        $0 is GestureClick
+                            && gtk_gesture_single_get_button($0.opaquePointer) == GDK_BUTTON_PRIMARY
+                    } as! GestureClick
+                gesture.pressed = { _, nPress, _, _ in
+                    guard environment.isEnabled, nPress == 1 else {
+                        return
+                    }
+                    action()
+                }
+            case .secondary:
+                let gesture =
+                    tapGestureTarget.eventControllers.first {
+                        $0 is GestureClick &&
+                            (gtk_gesture_single_get_button($0.opaquePointer) ==
+                                GDK_BUTTON_SECONDARY)
+                    } as! GestureClick
+                gesture.pressed = { _, nPress, _, _ in
+                    guard environment.isEnabled, nPress == 1 else {
+                        return
+                    }
+                    action()
+                }
+            case .longPress:
+                let gesture =
+                    tapGestureTarget.eventControllers.lazy.compactMap { $0 as? GestureLongPress }
+                        .first!
+                gesture.pressed = { _, _, _ in
+                    guard environment.isEnabled else {
+                        return
+                    }
+                    action()
+                }
+        }
+    }
+}
