+++ Sources/GtkBackend/Features/TextFields.swift
@@ -1,0 +1,41 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.TextFields {
+    public func createTextField() -> Widget {
+        Entry()
+    }
+
+    public func updateTextField(
+        _ textField: Widget,
+        placeholder: String,
+        environment: EnvironmentValues,
+        onChange: @escaping (String) -> Void,
+        onSubmit: @escaping () -> Void
+    ) {
+        let textField = textField as! Entry
+        textField.sensitive = environment.isEnabled
+        textField.placeholderText = placeholder
+        textField.changed = { widget in
+            onChange(widget.text)
+        }
+        textField.activate = { _ in
+            onSubmit()
+        }
+
+        textField.css.clear()
+        textField.css.set(properties: Self.cssProperties(for: environment, isControl: true))
+    }
+
+    public func setContent(ofTextField textField: Widget, to content: String) {
+        let textField = textField as! Entry
+        textField.withBlockedSignal(named: "changed") {
+            textField.text = content
+        }
+    }
+
+    public func getContent(ofTextField textField: Widget) -> String {
+        (textField as! Entry).text
+    }
+}
