+++ Sources/GtkBackend/Features/ToggleButtons.swift
@@ -1,0 +1,35 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.ToggleButtons {
+    public func createToggle() -> Widget {
+        return ToggleButton()
+    }
+
+    public func updateToggle(
+        _ toggle: Widget,
+        label: String,
+        environment: EnvironmentValues,
+        onChange: @escaping (Bool) -> Void
+    ) {
+        let toggle = toggle as! Gtk.ToggleButton
+        toggle.sensitive = environment.isEnabled
+        toggle.label = label
+        toggle.toggled = { widget in
+            onChange(widget.active)
+        }
+        toggle.css.clear()
+        // This is a control, but we set isControl to false anyway because isControl overrides
+        // the button background and makes the on and off states of the toggle look identical.
+        toggle.css.set(properties: Self.cssProperties(for: environment, isControl: false))
+    }
+
+    public func setState(ofToggle toggle: Widget, to state: Bool) {
+        let toggle = toggle as! Gtk.ToggleButton
+
+        toggle.withBlockedSignal(named: "toggled") {
+            toggle.active = state
+        }
+    }
+}
