+++ Sources/GtkBackend/Features/Tooltips.swift
@@ -1,0 +1,14 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.Tooltips {
+    public func createTooltipContainer(wrapping child: Widget) -> Widget {
+        TooltipContainer(child)
+    }
+
+    public func updateTooltipContainer(_ widget: Widget, tooltip: String) {
+        let widget = widget as! TooltipContainer
+        widget.setTooltip(text: tooltip)
+    }
+}
