+++ Sources/GtkBackend/Features/Widgets.swift
@@ -1,0 +1,29 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.Widgets {
+    public typealias Widget = Gtk.Widget
+
+    public var defaultPaddingAmount: Int { 10 }
+    
+    public func show(widget: Widget) {
+        widget.show()
+    }
+
+    public func tag(widget: Widget, as tag: String) {
+        widget.tag(as: tag)
+    }
+
+    public func naturalSize(of widget: Widget) -> SIMD2<Int> {
+        let currentSize = widget.getSizeRequest()
+        widget.setSizeRequest(width: -1, height: -1)
+        let (width, height) = widget.getNaturalSize()
+        widget.setSizeRequest(width: currentSize.width, height: currentSize.height)
+        return SIMD2(width, height)
+    }
+
+    public func setSize(of widget: Widget, to size: SIMD2<Int>) {
+        widget.setSizeRequest(width: size.x, height: size.y)
+    }
+}
