+++ Sources/GtkBackend/Features/WindowClosing.swift
@@ -1,0 +1,19 @@
+import CGtk
+import Gtk
+@_spi(Backends) import SwiftCrossUI
+
+extension GtkBackend: BackendFeatures.WindowClosing {
+    public func close(window: Window) {
+        window.close()
+    }
+
+    public func setCloseHandler(
+        ofWindow window: Window,
+        to action: @escaping () -> Void
+    ) {
+        window.onCloseRequest = { _ in
+            action()
+            window.destroy()
+        }
+    }
+}
