+++ Sources/GtkBackend/Helpers/CustomListBox.swift
@@ -1,0 +1,6 @@
+import Gtk
+
+class CustomListBox: ListBox {
+    var cachedSelection: Int? = nil
+    var cachedRowCount = 0
+}
