+++ Sources/GtkBackend/Helpers/Extensions.swift
@@ -1,0 +1,6 @@
+extension UnsafeMutablePointer {
+    func cast<T>() -> UnsafeMutablePointer<T> {
+        let pointer = UnsafeRawPointer(self).bindMemory(to: T.self, capacity: 1)
+        return UnsafeMutablePointer<T>(mutating: pointer)
+    }
+}
