+++ Sources/GtkBackend/Helpers/GtkCustomButton.swift
@@ -1,0 +1,59 @@
+import Gtk
+import CGtk
+@_spi(Backends) import SwiftCrossUI
+
+final class GtkCustomButton: Gtk.Button {
+    var buttonStyle: ButtonStyle.Kind = .bordered {
+        willSet {
+            buttonStyle.removeClass(from: self)
+        }
+        didSet {
+            buttonStyle.setClass(on: self)
+        }
+    }
+
+    init() {
+        super.init(gtk_button_new())
+
+        gtk_widget_add_css_class(widgetPointer, "customButton")
+    }
+
+    @MainActor
+    func loadCSS(environment: EnvironmentValues) {
+        let backgroundColor = GtkBackend.controlBackgroundColor(for: environment)
+        cssProvider.loadCss(from: """
+                button.customButton {
+                    min-width: 0px;
+                    min-height: 0px;
+                    padding: 0px;
+                    background: \(CSSProperty.rgba(backgroundColor));
+                    border: none;
+                    box-shadow: none;
+                }
+
+                button.customButton.flat:active,
+                button.customButton.flat.keyboard-activating {
+                    opacity: 0.80;
+                }
+
+                button.customButton.flat {
+                    background: transparent;
+                }
+
+                button.customButton.flat:focus {
+                    border-radius: 0px;
+                }
+
+                button.customButton.flat:disabled {
+                    opacity: 0.5;
+                }
+            """)
+        // Why 50% disabled opacity was chosen:
+        // https://gnome.pages.gitlab.gnome.org/libadwaita/doc/main/css-variables.html#opacity
+        // (switch to the variable when we have adwaita)
+
+        // Why 80% for active(pressed) was chosen:
+        // A pressed SwiftUI .plain button looks visually the same as
+        // a not pressed one at 0.8 opacity.
+    }
+}
